requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
trafilatura==1.6.1
duckduckgo-search==3.8.3
Pillow==11.0.0
python-telegram-bot==21.0.1
tenacity==8.2.3
requests==2.31.0
